Removed trailing slashes and added div
After getting feedback from Thinkful they told me I didn't pass because of: "HTML should be semantic and valid." After looking at the Validation report multiple times I have chiseled the solution down to the point that I have "trailing slashes" (which were not put in my me; slashes were defaulted code from Thinkful), and I believe one section of code was not inside a semantic container. 

I have erased all trailing slashes for sake of the validator not noting them, and added a div container to the "About" portion of the site to improve semantics.

If this is incorrect I do not know what else to fix as everything looks good to me and it matches the example design.